agent_core/requirements.txt
fastapi
sqlalchemy
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
uvicorn
psycopg2-binary
google-generativeai
python-dotenv
python-telegram-bot
